reworks the function structure
