+++ package.json
@@ -10,9 +10,9 @@
     "build": "rimraf __build__ && babel ./src -d ./__build__ --copy-files --ignore ./**/__tests__",
     "postbuild": "cp .env __build__",
     "predeploy:api:graphEndpoint": "npm run build",
-    "deploy:api:graphEndpoint": "functions-framework --target=graphEndpoint --port=8000 --source=./__build__",
+    "deploy:api:graphEndpoint": "functions-framework --target=handler --port=8000 --source=./__build__/graphql",
     "predeploy:api:fileUploadEndpoint": "npm run build",
-    "deploy:api:fileUploadEndpoint": "functions-framework --target=fileUploadEndpoint --port=7999 --source=./__build__",
+    "deploy:api:fileUploadEndpoint": "functions-framework --target=handler --port=7999 --source=./__build__/fileUpload",
     "start:watch:api:graphApi": "nodemon -e env,js,json,graphql --watch src --ignore '*.test.js' --exec npm run deploy:api:graphEndpoint",
     "start:watch:api:fileUploadApi": "nodemon -e env,js,json,graphql --watch src --ignore '*.test.js' --exec npm run deploy:api:fileUploadEndpoint",
     "prestart:watch:this": "sleep 5",
