--- src/index.js
@@ -1,5 +1,0 @@
-import graphApi from './graphApi';
-import fileUploadApi from './fileUploadApi';
-
-export const graphEndpoint = graphApi;
-export const fileUploadEndpoint = fileUploadApi;
